Function numcodecs_uniform_noise::add_uniform_noise

pub fn add_uniform_noise<T: Float, D: Dimension>(
    data: CowArray<'_, T, D>,
    scale: T,
    seed: u64,
) -> Array<T, D>
where
    Open01: Distribution<T>,

Uniform noise codec which adds U(-scale/2, scale/2) uniform random noise to the input on encoding and passes through the input unchanged during decoding.

This codec first hashes the input and its shape to then seed a pseudo-random number generator that generates the uniform noise. Therefore, encoding the same data with the same seed will produce the same noise and thus the same encoded data.